+++ packages/falcon-ecommerce-uikit/src/Footer/index.ts
@@ -1,2 +1,1 @@
 export * from './FooterSections';
-export * from './Copyright';
