+++ packages/falcon-ui-kit/src/Footer/FooterBanner.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Box, themed } from '@deity/falcon-ui';
+
+export const FooterBanner = themed({
+  tag: Box,
+  defaultTheme: {
+    footerBanner: {
+      display: 'grid',
+      gridGap: 'sm',
+      m: 'none',
+      p: 'none',
+      color: 'secondaryText',
+      bgFullWidth: 'secondary',
+      css: {
+        textAlign: 'center'
+      }
+    }
+  }
+});
