+++ packages/falcon-ui-kit/src/Footer/index.ts
@@ -1,1 +1,3 @@
 export * from './Footer';
+export * from './FooterBanner';
+export * from './Copyright';
